decouple conversion from main.js
